requirements.txt
tabula-py==2.9.3
pandas==2.2.3
requests==2.32.3
beautifulsoup4==4.12.3
openpyxl==3.1.5
